Make pixel seed/color persistent and add tunnel edge effect
Any pixel is now able to remember it's seed aka color table index after it moves, after a lot of minor refactoring. Was already added to the rendering side, but the CPU side didn't persistent the pixel seed.

This is now utilized to create the tunnel edge effect which makes the edges of the tunnels you dig slightly darker/more grovelled up looking.